Updated the Creating GradeBookType Enum task and tests.
