--- GradeBookTests/GradeBookUserInterfaceTests.cs
@@ -1,6 +1,0 @@
-﻿namespace GradeBookTests
-{
-    public class GradeBookUserInterfaceTests
-    {
-    }
-}
