+++ GradeBookTests/RankedGradingTests.cs
@@ -1,0 +1,56 @@
+﻿using System;
+using System.IO;
+using System.Linq;
+using Xunit;
+
+namespace GradeBookTests
+{
+    /// <summary>
+    ///     This class contains all tests related to the implimentation of adding support for Ranked Grading.
+    ///     Note: Do not use these tests as example of good testing practices, due to the nature of how Pluralsight projects work
+    ///     we have to create tests against code that doesn't exist and changes implimentation, due to this tests are fragile,
+    ///     hard to maintain, and don't don't adhere to the "test just one thing" practice commonly used in production tests.
+    /// </summary>
+    public class RankedGradingTests
+    {
+        /// <summary>
+        ///     Tests all requirements for creating the `GradeBookType` enum.
+        /// </summary>
+        [Fact]
+        [Trait("Category", "CreateGradeBookTypeEnum")]
+        public void GradeBookTypeEnumTests()
+        {
+            // Test to make sure `GradeBookType.cs` is in the `Enums` directory.
+            var filePath = ".." + Path.DirectorySeparatorChar + ".." + Path.DirectorySeparatorChar + ".." + Path.DirectorySeparatorChar + ".." + Path.DirectorySeparatorChar + "GradeBook" + Path.DirectorySeparatorChar + "Enums" + Path.DirectorySeparatorChar + "GradeBookType.cs";
+            Assert.True(File.Exists(filePath), "`GradeBookType.cs` was not found in the `Enums` folder.");
+
+            // Test to make sure the enum `GradeBookType` exists in the `GradeBooks.Enums` namespace.
+            var gradebookEnum = (from assembly in AppDomain.CurrentDomain.GetAssemblies()
+                                 from type in assembly.GetTypes()
+                                 where type.FullName == "GradeBook.Enums.GradeBookType"
+                                 select type).FirstOrDefault();
+            Assert.True(gradebookEnum != null, "`GradeBook.Enums.GradeBookType` wasn't found in the `GradeBooks.Enums` namespace.");
+
+            // Test to make sure the enum `GradeBookType` is public.
+            Assert.True(gradebookEnum.IsPublic, "`GradeBook.Enums.GradeBookType` exists, but isn't `public`.");
+
+            // Test to make sure that `GradeBookType` is an enum not a class
+            Assert.True(gradebookEnum.IsEnum, "`GradeBook.Enums.GradeBookType` exists, but isn't an enum.");
+
+            // Test that `GradeBookType` contains the value `Standard`
+            Assert.True(Enum.Parse(gradebookEnum, "Standard", true) != null, "`GradeBook.Enums.GradeBookType` doesn't contain the value `Standard`.");
+
+            // Test that `GradeBookType` contains the value `Ranked`
+            Assert.True(Enum.Parse(gradebookEnum, "Ranked", true) != null, "`GradeBook.Enums.GradeBookType` doesn't contain the value `Ranked`.");
+
+            // Test that `GradeBookType` contains the value `ESNU`
+            Assert.True(Enum.Parse(gradebookEnum, "ESNU", true) != null, "`GradeBook.Enums.GradeBookType` doesn't contain the value `ESNU`.");
+
+            // Test that `GradeBookType` contains the value `OneToFour`
+            Assert.True(Enum.Parse(gradebookEnum, "OneToFour", true) != null, "`GradeBook.Enums.GradeBookType` doesn't contain the value `OneToFour`.");
+
+            // Test that `GradeBookType` contains the value `SixPoint`
+            Assert.True(Enum.Parse(gradebookEnum, "SixPoint", true) != null, "`GradeBook.Enums.GradeBookType` doesn't contain the value `SixPoint`.");
+        }
+    }
+}
